Create README.md
This project is the Airgead Bank app that detects bank deposits, balances and more.
I did very well on putting the code together but I do forget to put in-lie comments, so that could be a little better as
I code. The comments on the project would help the user navigate better as they look through it. 
There were a few challenging parts to this project, one being the the for loop. This loop had me a little confused but I got it after a little research.
I used W3Schools to help me out with this problem. 
The menu displayed can be used in other projects, with a little moving around. I have done menus in other proects before.
Having my code in the right area and having commments built around the code.